+++ clinica/pipelines/cli.py
@@ -1,5 +1,7 @@
 import click
 
+from clinica.pydra.t1_linear import t1_linear_cli as pydra_t1_linear_cli
+
 from .deeplearning_prepare_data import (
     deeplearning_prepare_data_cli as deeplearning_prepare_data_cli,
 )
@@ -29,7 +31,6 @@ from .t1_volume_existing_template import t1_volume_existing_template_cli
 from .t1_volume_parcellation import t1_volume_parcellation_cli
 from .t1_volume_register_dartel import t1_volume_register_dartel_cli
 from .t1_volume_tissue_segmentation import t1_volume_tissue_segmentation_cli
-from .pydra_t1_linear import t1_linear_cli
 
 
 class RegistrationOrderGroup(click.Group):
@@ -74,7 +75,7 @@ cli.add_command(t1_freesurfer_template_cli.cli)
 cli.add_command(t1_freesurfer_longitudinal_correction_cli.cli)
 cli.add_command(deeplearning_prepare_data_cli.cli)
 
-cli.add_command(t1_linear_cli.cli)  # pydra option
+cli.add_command(pydra_t1_linear_cli.cli)  # pydra option
 
 if __name__ == "__main__":
     cli()
